+++ changelog/mediation/mobileads-mintegral/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 7.5.6.0
+
+### Updated
+
+* Added support for Mintegral SDK version 7.5.6
+* Updated minimum supported Mintegral SDK version to 7.5.6
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.4.0
+
 ## Version 7.5.1.0
 
 ### Updated